#2 Расчет ширины View, расчет ширины ячеек, отрисовка координатной сетки, отрисовка усредненных значений

+++ app/src/main/java/ru/dmisb/testchart/utils/Ext.kt
@@ -5,12 +5,12 @@ import android.util.TypedValue
 import androidx.annotation.ColorRes
 import androidx.core.content.ContextCompat
 
-fun Context.dpToPx(value: Int) : Int =
+fun Context.dpToPx(value: Float) : Float =
     TypedValue.applyDimension(
         TypedValue.COMPLEX_UNIT_DIP,
-        value.toFloat(),
+        value,
         this.resources.displayMetrics
-    ).toInt()
+    )
 
 fun Context.color(@ColorRes colorRes: Int) : Int =
     ContextCompat.getColor(this, colorRes)
